ZeldaDS:
* All data now loaded from file system (no longer linked into rom). This includes: foreground and background tiles, item tiles, and character sprites.
** Added ResourceMgr to manage pools of memory for large resource allocations
** Added GameResources.h for game-specific resource enums

* Improved some batch files to simplify using them

* build.bat + makefile: now emits the rom section sizes for the arm9 and arm7 elf files post-build

[git-p4: depot-paths = "//depot/NDS/ZeldaDS/": change = 220]

+++ Game/ZeldaDS/arm9/source/gslib/Hw/BackgroundLayer.cpp
@@ -66,7 +66,7 @@ void BackgroundLayer::InitTiled(int bgId, uint8 metaTileSizeX, uint8 metaTileSiz
 	memset(mpTileMap, 0, (HwBgNumTilesX * HwBgNumTilesY * tileIndexSizeBytes)); //@TODO: use dmaCopy?
 }
 
-void BackgroundLayer::LoadTilesImage(const uint8* pImage, uint16 sizeBytes)
+void BackgroundLayer::LoadTilesImage(const uint16* pImage, uint16 sizeBytes)
 {
 	memcpy(bgGetGfxPtr(mBgId), pImage, sizeBytes); //@TODO: use dmaCopy?
 }
