+++ Game/ZeldaDS/arm9/source/gslib/Hw/BackgroundLayer.h
@@ -13,7 +13,7 @@ public:
 
 	// Tile layer functions
 	void InitTiled(int bgId, uint8 metaTileSizeX, uint8 metaTileSizeY);
-	void LoadTilesImage(const uint8* pImage, uint16 sizeBytes);
+	void LoadTilesImage(const uint16* pImage, uint16 sizeBytes);
 	void DrawTile(uint16 metaTileIndex, uint16 metaTileMapX, uint16 metaTileMapY);
 
 	// Returns pointer to the tiles image buffer
